refactor: Move types file to separate directory
