+++ packages/frontend/src/Daily/SelectRoutine/RoutineItem.tsx
@@ -4,9 +4,9 @@ import { useQueryClient } from 'react-query';
 import { useRecoilValue } from 'recoil';
 
 import { createWorkoutsByRoutine } from '../../apis/workoutsAPI';
+import { IRoutine } from '../../types';
 import { getDateString } from '../../utils/getDateString';
 import { dateState } from '../states';
-import { IRoutine } from '../types';
 import ExerciseItem from './ExerciseItem';
 
 const TextAlignLeftCardContent = styled(CardContent)({
