+++ packages/frontend/src/hooks/useRoutines.ts
@@ -1,7 +1,7 @@
 import { useQuery, UseQueryResult } from "react-query";
 
 import { getRoutines } from "../apis/workoutsAPI";
-import { IRoutine } from "../Daily/types";
+import { IRoutine } from "../types";
 
 export function useRoutines(): UseQueryResult<IRoutine[], Error> {
   return useQuery<IRoutine[], Error>(`routines`, () => getRoutines());
